Spin out HCSShim building code
